feat(reapply): allow reapplying vaporization
Refactor the mods into their own module
Also, move vaporization to its own function so it can be reapplied
by pressing any button except escape
